+++ scripts/beaker/mitchish7-launch.sh
@@ -5,10 +5,10 @@ set -ex
 NUM_NODES=8
 
 gantry run \
-  --workspace ai2/shanea \
-  --task-name mitchish7-find-spike \
-  --description "Look for loss & grad spikes in OLMo 1.7 7B run" \
-  --priority high \
+  --workspace ai2/dirkg \
+  --task-name mitchish7 \
+  --description "OLMo medium - 7B" \
+  --priority normal \
   --beaker-image shanea/olmo-torch2.2-gantry \
   --cluster ai2/jupiter-cirrascale \
   --gpus 8 \
@@ -17,6 +17,8 @@ gantry run \
   --host-networking \
   --budget ai2/oe-training \
   --no-nfs \
+  --propagate-failure \
+  --synchronized-start-timeout 10m \
   --env LOG_FILTER_TYPE=local_rank0_only \
   --env OMP_NUM_THREADS=8 \
   --env OLMO_TASK=model \
@@ -34,4 +36,4 @@ gantry run \
   --venv base \
   --yes \
   --timeout=-1 \
-  -- /bin/bash -c "scripts/beaker/mitchish7.sh \$BEAKER_LEADER_REPLICA_HOSTNAME ${NUM_NODES}"
+  -- /bin/bash -c "scripts/beaker/mitchish7.sh \$BEAKER_LEADER_REPLICA_HOSTNAME ${NUM_NODES} \$BEAKER_REPLICA_RANK"